requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
aiohttp
pydantic
typing-extensions